+++ src/resources/defaultConfig.ini
@@ -198,9 +198,9 @@ SpeedTest           = 100
 [Video]
 ; Screen rendering mode.
 ; OpenGL ES 3.2 (Android only, default)
-; OpenGL 3.2 (desktop only, default)
+; OpenGL 3.3 (desktop only, default)
 ; Vulkan 1.3 (desktop only, experimental)
-RenderMode        = OpenGL 3.2
+RenderMode        = OpenGL 3.3
 ; Game native width and height.
 ; Recommended settings are:
 ; 640x480   Standard definition 4:3
